Move set_columns from argparse to console
