+++ pytoolbox/console.py
@@ -5,9 +5,21 @@ from typing import Any, TextIO
 import atexit
 import code
 import os
+import shutil
 import sys
 
-__all__ = ['confirm', 'choice', 'print_error', 'progress_bar', 'shell', 'toggle_colors']
+__all__ = [
+    'confirm',
+    'choice',
+    'print_error',
+    'progress_bar',
+    'shell',
+    'set_columns',
+    'toggle_colors'
+]
+
+
+# Input --------------------------------------------------------------------------------------------
 
 
 def confirm(
@@ -82,6 +94,9 @@ def choice(
         stream.write(f'Please choose between {choices_string}.{os.linesep}')
 
 
+# Output -------------------------------------------------------------------------------------------
+
+
 def print_error(message: str, exit_code: int | None = 1, stream: TextIO = sys.stderr) -> None:
     """
     Print an error message and exit if `exit_code` is not None.
@@ -132,6 +147,9 @@ def progress_bar(
         stream.flush()
 
 
+# Interactivity ------------------------------------------------------------------------------------
+
+
 def shell(
     banner: str | None = None,
     *,
@@ -157,6 +175,19 @@ def shell(
             pass
         atexit.register(readline.write_history_file, history_filename)
     code.interact(banner=banner, local=imported)
+
+
+# Configuration ------------------------------------------------------------------------------------
+
+
+def set_columns(value: int | None = None, *, default: int = 120) -> int:
+    if value is None:
+        try:
+            value = shutil.get_terminal_size().columns
+        except AttributeError:
+            value = default
+    os.environ['COLUMNS'] = str(value)
+    return value
 
 
 def toggle_colors(
